requirements.txt
gradio>=4.0.0
torch>=2.0.0
transformers>=4.30.0
diffusers>=0.21.0
accelerate>=0.21.0
safetensors>=0.3.1
einops>=0.6.1
numpy>=1.24.0
pillow>=9.4.0
